+++ src/wai/common/adams/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package for functionality for inter-operating with ADAMS.
+"""