+++ src/wai/common/adams/imaging/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package for inter-operating with the adams-imaging module.
+"""