+++ src/wai/common/adams/imaging/locateobjects/__init__.py
@@ -1,0 +1,5 @@
+"""
+Package for working with located objects in image classification/identification tasks.
+"""
+from ._LocatedObject import LocatedObject
+from ._LocatedObjects import LocatedObjects