+++ src/wai/common/adams/imaging/locateobjects/constants.py
@@ -1,0 +1,29 @@
+"""
+Module for constants relating to locating objects.
+"""
+# The key for the X location
+KEY_X: str = "x"
+
+# The key for the Y location
+KEY_Y: str = "y"
+
+# The key for the width
+KEY_WIDTH: str = "width"
+
+# The key for the height
+KEY_HEIGHT: str = "height"
+
+# The key for the location
+KEY_LOCATION: str = "location"
+
+# The key for the Xs of the polygon in the meta-data (comma-separated list)
+KEY_POLY_X: str = "poly_x"
+
+# The key for the Ys of the polygon in the meta-data (comma-separated list)
+KEY_POLY_Y: str = "poly_y"
+
+# The key for the overall count
+KEY_COUNT: str = "count"
+
+# The key for the index of a group
+KEY_INDEX: str = "index"